refactor: Profile画面のコンポーネント分割、命名整理
