+++ src/screens/profile/components/index.ts
@@ -1,2 +1,2 @@
-export * from './Avatar'
-export * from './Card'
+export * from './Nameplate'
+export * from './AccountList'
