+++ src/screens/profile/index.tsx
@@ -1,8 +1,6 @@
 import { StatusBar } from 'expo-status-bar'
-import { ImageBackground, ScrollView, StyleSheet, Text, View } from 'react-native'
-import { Avatar, Card } from './components'
-import { cards } from '#/constants/cards'
-import { COLORS } from '#/constants/environment'
+import { ImageBackground, ScrollView, StyleSheet, View } from 'react-native'
+import { AccountList, Nameplate } from './components'
 import { isLandscape } from '#/utils'
 
 export const ProfileScreen = () => {
@@ -11,17 +9,10 @@ export const ProfileScreen = () => {
       <ImageBackground source={require('#assets/beer.jpg')} style={styles.background}>
         <StatusBar style="auto" />
         <ScrollView style={styles.scroll}>
-          <View style={styles.headerContainer}>
-            <Avatar />
-            <Text style={styles.title}>asmz</Text>
-            <Text style={styles.description}>Akira Shimizu</Text>
-            <Text style={styles.description}>純東北産麦芽系エンジニア</Text>
-          </View>
-          <View style={styles.listContainer}>
-            {cards.map((item, index) => (
-              <Card key={index} card={item} />
-            ))}
-          </View>
+          {/* プロフィール情報 */}
+          <Nameplate />
+          {/* 各種サービスアカウントリスト */}
+          <AccountList />
         </ScrollView>
       </ImageBackground>
     </View>
@@ -44,23 +35,6 @@ const styles = StyleSheet.create({
   scroll: {
     paddingHorizontal: isLandscape() ? 250 : 0,
     paddingBottom: 30,
-  },
-  headerContainer: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingTop: 150,
-    paddingBottom: 50,
-  },
-  title: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    paddingTop: 20,
-    color: COLORS.text,
-  },
-  description: {
-    fontSize: 16,
-    padding: 20,
-    color: COLORS.text,
   },
   listContainer: {
     alignItems: 'center',
